requirements.txt
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
pyyaml>=6.0
loguru>=0.7.0
tqdm>=4.65.0

wandb>=0.16.0
scipy>=1.11.0

pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.0
ruff>=0.1.0
mypy>=1.5.0

torch>=2.1.0
wandb>=0.16.0
tensorboard>=2.15.0
